+++ assets/app/app/visualizer/validator.js
@@ -160,6 +160,8 @@ angular.module('a2.speciesValidator', ['a2.utils', 'a2.infotags'])
                     if (val[0] == 1 || val[1] > 0 || val[2] > 0) {
                         val_options[1].showDropdown = (val[1] > 0 || val[2] > 0)
                         return val_options[1];
+                    } if (val[0] == null && val[1] == 0 && val[2] == 0) {
+                        return;
                     } else {
                         return val_options[2];
                     }